# Add C acceleration for melting temperature calculations