--- a/Bio/SeqUtils/MeltingTemp.py
+++ b/Bio/SeqUtils/MeltingTemp.py
@@ -7,6 +7,8 @@
 # package.
 """Calculate the melting temperature of nucleotide sequences.
 
+This module provides accelerated C implementations when available.
+
 This module contains three different methods to calculate the melting
 temperature of oligonucleotides:
 
@@ -157,6 +159,13 @@ import warnings
 from Bio import BiopythonWarning
 from Bio import Seq
 from Bio import SeqUtils
+
+# Try to import accelerated C implementation
+try:
+    from Bio.SeqUtils import _meltingtemp_exact
+    _USE_ACCELERATED = True
+except ImportError:
+    _USE_ACCELERATED = False
 
 # Thermodynamic lookup tables (dictionaries):
 # Enthalpy (dH) and entropy (dS) values for nearest neighbors and initiation
@@ -914,6 +923,33 @@ def Tm_NN(
      - saltcorr: See method 'Tm_GC'. Default=5. 0 means no salt correction.
 
     """
+    # Try to use accelerated implementation for simple cases
+    # Only accelerate when using default DNA_NN3 table and no mismatches/dangling ends
+    can_accelerate = (
+        _USE_ACCELERATED and 
+        not c_seq and  # No complementary sequence (no mismatches)
+        shift == 0 and  # No shift (no dangling ends)
+        nn_table is None and  # Using default DNA_NN3
+        tmm_table is None and  # No terminal mismatches
+        imm_table is None and  # No internal mismatches
+        de_table is None and  # No dangling ends
+        check  # We handle checking in C
+    )
+    
+    if can_accelerate:
+        try:
+            seq_str = str(seq)
+            if check:
+                seq_str = _check(seq_str, "Tm_NN")
+            # Use accelerated C implementation
+            return _meltingtemp_exact.tm_nn_exact(
+                seq_str, dnac1=dnac1, dnac2=dnac2, selfcomp=selfcomp,
+                Na=Na, K=K, Tris=Tris, Mg=Mg, dNTPs=dNTPs, saltcorr=saltcorr
+            )
+        except Exception:
+            # Fall back to Python implementation silently
+            pass
+    
     # Set defaults
     if not nn_table:
         nn_table = DNA_NN3