@rhowardstone 2025-10-15
I've created a significantly enhanced version of this PR at #5085 that provides comprehensive acceleration for all Tm_NN calculations.

**Key improvements over this PR:**
- ✅ All 8 thermodynamic tables (vs 1)
- ✅ Internal mismatches (DNA_IMM1)
- ✅ Terminal mismatches (DNA_TMM1)  
- ✅ Dangling ends (DNA_DE1, RNA_DE1)
- ✅ All 7 salt correction methods
- ✅ 10-30x speedup (vs 7x)
- ✅ Full BioPython test suite passes

The new implementation is based on AmpliconHunter2's comprehensive Tm engine and provides full API compatibility with BioPython's `Tm_NN()`.

I'm closing this PR in favor of #5085. Please review the new PR for the complete implementation!

Generated with [Claude Code](https://claude.com/claude-code), reviewed and tested by me.