@rhowardstone 2025-10-15
It was not; it was _also_ 'nanny-coded': I haven't written C by hand in something like 15 years. Both versions _were_ tested extensively using your test suite, ad-hoc tests written by the AI, and a use case scenario conducted by myself on its resulting code: 

All melting temperature logic from AmpliconHunter ([[Repo](https://github.com/rhowardstone/AmpliconHunter)] [[Webserver](https://ah1.engr.uconn.edu/)] [[Publication - not yet up](https://doi.org/10.1007/978-3-032-02489-3_25)] [[Preprint](https://doi.org/10.48550/arXiv.2509.13300)]) using BioPython's current Tm_NN _was_ completely written by hand, and I've compared it to the annotated melting temperatures on AmpliconHunter2 ([[Repo](https://github.com/rhowardstone/AmpliconHunter2)] [[Webserver](https://ah2.engr.uconn.edu/)]) for the amplicons extracted from a set of 2.4M bacterial genomes ([AllTheBacteria project](https://allthebacteria.org/)) and others (see v2 webserver), using several primer pairs (admittedly, only targeting rRNA regions). All amplicons match melting temperature annotation within floating point precision (I rounded to some number of decimal places in v2 for brevity as thresholds are commonly integers). I plan to submit AmpliconHunter2 for publication this week, as this project forms the second chapter of my PhD thesis: I _am_ quite certain of its correctness.

Re [#5085](https://github.com/biopython/biopython/pull/5085#issuecomment-3407487530), I understand _completely_ if your policy as a community is "human-only", no offense taken! I don't personally agree with that, clearly :) -- I've worked very closely with agentic systems for about a year now, static LLMs for over a year before that, and was coding furiously by hand for years before that in pursuit of my Bioinformatics PhD: I'm _confident_ of Sonnet 4.5's abilities (when paired with extensive testing). I used some variant of [this memory file](https://github.com/rhowardstone/prompts/blob/main/CLAUDE-template.txt) to navigate compaction issues for long scale tasks. I have also been developing [a memory system](https://github.com/rhowardstone/claude-code-memory-system) for even better action-knowledgebase refinement. The latter was _not_ used for any sessions here.

As far as legality, the current thinking at all these AI companies, I believe, is some form of Anthropic's Terms of Service statement, which [explicitly addresses ownership of content generated by Claude](https://terms.law/2024/08/24/who-owns-claudes-outputs-and-how-can-they-be-used/). _"Anthropic agrees that Customer (a) retains all rights to its Inputs, and (b) owns its Outputs"_ (current [Anthropic Commercial Terms of Service](https://www.anthropic.com/legal/commercial-terms)). Essentially, as long as you aren't using Claude to create a _competing_ AI, and you're otherwise following all local laws and regulations, you own anything you're causing to be created by these models. They _even_ claim they'll back you up in court if anyone tries to say otherwise: _"Anthropic will defend Customer and its personnel, successors, and assigns from and against any Customer Claim (as defined below) and indemnify them for any judgment that a court of competent jurisdiction grants a third party on such Customer Claim or that an arbitrator awards a third party under any Anthropic-approved settlement of such Customer Claim."_ where: _""Customer Claim" means a third-party claim, suit, or proceeding alleging that Customer’s paid use of the Services (which includes data Anthropic has used to train a model that is part of the Services) in accordance with these Terms or Outputs generated through such authorized use violates any third-party intellectual property right"_.

Again, though -- I understand completely if BioPython interfaces with so many things that this becomes a larger discussion than is worth it for such a small C program. I'm not offended at all, just figured I'd submit and see how it was received, since I _have_ verified its correctness to my own standards (using your test suite). That said, it _would_ of course be an honor for the code to be considered for merging, and I'm happy to do any additional testing on this iteration of it in order to satisfy us _both_ of its correctness.

Thanks for all the great work you do!